+++ Changelog.md
@@ -4,6 +4,8 @@ Expect active development and potentially significant breaking changes in the `0
 
 ### vNext
 
+### v0.7.0
+
 #### Breaking
 ```js
 // old